+++ src/main/java/main/model/db/dao/project/StepResultDao.java
@@ -1,0 +1,13 @@
+package main.model.db.dao.project;
+
+import main.model.db.dao.DAO;
+import main.model.dto.StepResultDto;
+
+public class StepResultDao extends DAO<StepResultDto> {
+    public StepResultDao() {
+        super(StepResultDto.class);
+        select = "{call SELECT_STEP_RESULT(?,?,?)}";
+        insert = "{call INSERT_STEP_RESULT(?,?,?,?,?,?,?,?,?,?,?,?)}";
+        remove = "{call REMOVE_STEP_RESULT(?)}";
+    }
+}