+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # CHANGELOG
 
-## 0.3.2 (unreleased)
+## 0.3.3 (unreleased)
+
+Features:
+  - Rename Import Token to API Token -> [View Issue](https://github.com/aquality-automation/aquality-tracking-ui/issues/23)
+  - Steps Feature -> [View Issue](https://github.com/aquality-automation/aquality-tracking-ui/issues/46)
+  - Regex Search by Errors for Import -> [View Issue](https://github.com/aquality-automation/aquality-tracking/issues/17)
+
+Bugfixes:
+  - Local Manager can see Local Permissions page -> [View Issue](https://github.com/aquality-automation/aquality-tracking-ui/issues/22)
+
+## 0.3.2 (2019-09-10)
 
 Features: